+++ crates/ntx-builder/src/actor/mod.rs
@@ -4,10 +4,12 @@ mod inflight_note;
 mod note_state;
 
 use std::sync::Arc;
+use std::time::Duration;
 
 use account_state::{NetworkAccountState, TransactionCandidate};
 use execute::NtxError;
 use futures::FutureExt;
+use miden_node_proto::clients::{Builder, ValidatorClient};
 use miden_node_proto::domain::account::NetworkAccountPrefix;
 use miden_node_proto::domain::mempool::MempoolEvent;
 use miden_node_utils::ErrorReport;
@@ -54,6 +56,8 @@ pub struct AccountActorContext {
     pub store: StoreClient,
     /// Address of the block producer gRPC server.
     pub block_producer_url: Url,
+    /// Address of the Validator server.
+    pub validator_url: Url,
     /// Address of the remote prover. If `None`, transactions will be proven locally, which is
     // undesirable due to the performance impact.
     pub tx_prover_url: Option<Url>,
@@ -153,7 +157,9 @@ pub struct AccountActor {
     mode: ActorMode,
     event_rx: mpsc::Receiver<Arc<MempoolEvent>>,
     cancel_token: CancellationToken,
+    // TODO(sergerad): Remove block producer when block proving moved to store.
     block_producer: BlockProducerClient,
+    validator: ValidatorClient,
     prover: Option<RemoteTransactionProver>,
     chain_state: Arc<RwLock<ChainState>>,
     script_cache: LruCache<Word, NoteScript>,
@@ -169,6 +175,13 @@ impl AccountActor {
         cancel_token: CancellationToken,
     ) -> Self {
         let block_producer = BlockProducerClient::new(actor_context.block_producer_url.clone());
+        let validator = Builder::new(actor_context.validator_url.clone())
+            .without_tls()
+            .with_timeout(Duration::from_secs(10))
+            .without_metadata_version()
+            .without_metadata_genesis()
+            .with_otel_context_injection()
+            .connect_lazy::<ValidatorClient>();
         let prover = actor_context.tx_prover_url.clone().map(RemoteTransactionProver::new);
         Self {
             origin,
@@ -177,6 +190,7 @@ impl AccountActor {
             event_rx,
             cancel_token,
             block_producer,
+            validator,
             prover,
             chain_state: actor_context.chain_state.clone(),
             script_cache: actor_context.script_cache.clone(),
@@ -276,6 +290,7 @@ impl AccountActor {
         // Execute the selected transaction.
         let context = execute::NtxContext::new(
             self.block_producer.clone(),
+            self.validator.clone(),
             self.prover.clone(),
             self.store.clone(),
             self.script_cache.clone(),