+++ crates/ntx-builder/src/builder.rs
@@ -73,6 +73,8 @@ pub struct NetworkTransactionBuilder {
     store_url: Url,
     /// Address of the block producer gRPC server.
     block_producer_url: Url,
+    /// Address of the Validator server.
+    validator_url: Url,
     /// Address of the remote prover. If `None`, transactions will be proven locally, which is
     /// undesirable due to the performance impact.
     tx_prover_url: Option<Url>,
@@ -95,6 +97,7 @@ impl NetworkTransactionBuilder {
     pub fn new(
         store_url: Url,
         block_producer_url: Url,
+        validator_url: Url,
         tx_prover_url: Option<Url>,
         ticker_interval: Duration,
         bp_checkpoint: Arc<Barrier>,
@@ -105,6 +108,7 @@ impl NetworkTransactionBuilder {
         Self {
             store_url,
             block_producer_url,
+            validator_url,
             tx_prover_url,
             ticker_interval,
             bp_checkpoint,
@@ -141,6 +145,7 @@ impl NetworkTransactionBuilder {
 
         let actor_context = AccountActorContext {
             block_producer_url: self.block_producer_url.clone(),
+            validator_url: self.validator_url.clone(),
             tx_prover_url: self.tx_prover_url.clone(),
             chain_state: chain_state.clone(),
             store: store.clone(),